+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Issue [#15](https://github.com/m-marini/qucomp/issues/15): Add function mapping state
 
+### Changed
+
+- Issue [#17](https://github.com/m-marini/qucomp/issues/17): Change state mapping function definition
+
 ### Fixed
 
 - Issue [#13](https://github.com/m-marini/qucomp/issues/13): Error on ccnot